+++ inc/admin/class-we-options.php
@@ -99,12 +99,22 @@ class WE_Options {
 	 * @return string
 	 */
 	public function setting_page_id() {
+		$page_option = array( '' => esc_html__( '- Disabled -', '%TEXTDOMAIN%' ) );
 		$pages = get_pages();
+
+		foreach ( $pages as $page ) {
+
+			if ( get_post_field( 'post_parent', $page->ID ) ) {
+				$page_option[ absint( $page->ID ) ] = '&nbsp;&nbsp;&nbsp; ' . sanitize_text_field( $page->post_title );
+			} else {
+				$page_option[ absint( $page->ID ) ] = sanitize_text_field( $page->post_title );
+			}
+		}
 		?>
 		<select name="wolf_events_settings[page_id]">
 			<option value="-1"><?php esc_html_e( 'Select a page...', '%TEXTDOMAIN%' ); ?></option>
-			<?php foreach ( $pages as $page ) : ?>
-				<option value="<?php echo absint( $page->ID ); ?>" <?php selected( absint( $page->ID ), get_option( '_wolf_events_page_id' ) ); ?>><?php echo sanitize_text_field( $page->post_title ); ?></option>
+			<?php foreach ( $page_option as $k => $v ) : ?>
+				<option value="<?php echo absint( $k ); ?>" <?php selected( absint( $k ), get_option( '_wolf_events_page_id' ) ); ?>><?php echo sanitize_text_field( $v ); ?></option>
 			<?php endforeach; ?>
 		</select>
 		<?php